requirements.txt
aiohttp<=3.8.1
pyrotgfork==2.1.32.15
python-dotenv<=0.21.0
tgcrypto<=1.2.3
pycryptodome==3.18.0
requests

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
